Initial commit
The first version of Steve the humanoid. The committed code is for the head using a time of flight sensor, two OLEDs and 5 microphones(that part is being restructured to only use one or two). The microphones were used as input for the eye movement but that part should be handled via face recognition.